middleware-service/clone_repo/requirements.txt
flask
GitPython
requests
